requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dotenv==0.9.9
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
razorpay==1.4.2
requests==2.32.3
setuptools==78.1.0
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.13.0
urllib3==2.4.0
Werkzeug==3.1.3
